Lint

+++ src/main/java/com/fengsheng/phase/SendPhaseStart.java
@@ -15,22 +15,20 @@ public record SendPhaseStart(Player player) implements Fsm {
     @Override
     public ResolveResult resolve() {
         Game game = player.getGame();
-        if (player.isAlive()) {
-            if (player.getCards().isEmpty()) {
-                log.info(player + "没有情报可传，输掉了游戏");
-                game.getDeck().discard(player.deleteAllMessageCards());
-                player.setLose(true);
-                player.setAlive(false);
-                for (Player p : game.getPlayers())
-                    p.notifyDying(player.location(), true);
-                Player alivePlayer = getOnlyOneAlivePlayer(game.getPlayers());
-                if (alivePlayer != null) {
-                    CheckKillerWin.onlyOneAliveWinner(game, alivePlayer);
-                    return new ResolveResult(null, false);
-                }
-                for (Player p : game.getPlayers())
-                    p.notifyDie(player.location());
+        if (player.isAlive() && player.getCards().isEmpty()) {
+            log.info(player + "没有情报可传，输掉了游戏");
+            game.getDeck().discard(player.deleteAllMessageCards());
+            player.setLose(true);
+            player.setAlive(false);
+            for (Player p : game.getPlayers())
+                p.notifyDying(player.location(), true);
+            Player alivePlayer = getOnlyOneAlivePlayer(game.getPlayers());
+            if (alivePlayer != null) {
+                CheckKillerWin.onlyOneAliveWinner(game, alivePlayer);
+                return new ResolveResult(null, false);
             }
+            for (Player p : game.getPlayers())
+                p.notifyDie(player.location());
         }
         if (!player.isAlive()) {
             return new ResolveResult(new NextTurn(player), true);
